Split webhook into separate package
