+++ pkg/registry/apis/provisioning/import.go
@@ -15,7 +15,7 @@ import (
 )
 
 type importConnector struct {
-	repoGetter RepoGetter
+	repoGetter Getter
 	client     *resources.ClientFactory
 	logger     *slog.Logger
 }
