+++ pkg/registry/apis/provisioning/register.go
@@ -7,6 +7,7 @@ import (
 	"encoding/hex"
 	"fmt"
 	"log/slog"
+	"net/http"
 	"net/url"
 	"path/filepath"
 	"slices"
@@ -15,6 +16,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -35,6 +37,7 @@ import (
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository/github"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/resources"
+	"github.com/grafana/grafana/pkg/registry/apis/provisioning/webhook"
 	"github.com/grafana/grafana/pkg/services/apiserver/builder"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/grafana/grafana/pkg/services/rendering"
@@ -48,7 +51,7 @@ const (
 
 var (
 	_ builder.APIGroupBuilder = (*ProvisioningAPIBuilder)(nil)
-	_ RepoGetter              = (*ProvisioningAPIBuilder)(nil)
+	_ Getter                  = (*ProvisioningAPIBuilder)(nil)
 )
 
 // This is used just so wire has something unique to return
@@ -224,6 +227,48 @@ func (b *ProvisioningAPIBuilder) GetRepository(ctx context.Context, name string)
 	return b.asRepository(ctx, obj)
 }
 
+func (b *ProvisioningAPIBuilder) GetWebhook(ctx context.Context, name string) (webhook.Webhook, error) {
+	repo, err := b.GetRepository(ctx, name)
+	if err != nil {
+		return nil, fmt.Errorf("get repository: %w", err)
+	}
+
+	namespace := repo.Config().GetNamespace()
+
+	if repo.Config().Spec.Type != provisioning.GitHubRepositoryType {
+		return nil, &apierrors.StatusError{
+			ErrStatus: v1.Status{
+				Message: fmt.Sprintf("webhook is not implemented for: %s", repo.Config().Spec.Type),
+				Code:    http.StatusNotImplemented,
+			},
+		}
+	}
+
+	dynamicClient, kinds, err := b.client.New(namespace)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create dynamic client: %w", err)
+	}
+
+	parser := resources.NewParser(repo.Config(), dynamicClient, kinds)
+	replicator := resources.NewReplicator(dynamicClient, parser, repo.Config().Spec.Folder)
+
+	linterFactory := lint.NewDashboardLinterFactory()
+	ghClient := b.ghFactory.New(ctx, repo.Config().Spec.GitHub.Token)
+	baseURL, err := url.Parse(b.urlProvider(namespace))
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse base url: %w", err)
+	}
+
+	return webhook.NewGithubWebhook(
+		repo,
+		replicator,
+		ghClient,
+		baseURL,
+		linterFactory,
+		b.renderer,
+	), nil
+}
+
 func (b *ProvisioningAPIBuilder) asRepository(ctx context.Context, obj runtime.Object) (repository.Repository, error) {
 	if obj == nil {
 		return nil, fmt.Errorf("missing repository object")
@@ -233,33 +278,17 @@ func (b *ProvisioningAPIBuilder) asRepository(ctx context.Context, obj runtime.O
 		return nil, fmt.Errorf("expected repository configuration")
 	}
 
-	namespace := r.GetNamespace()
-
 	switch r.Spec.Type {
 	case provisioning.LocalRepositoryType:
 		return repository.NewLocal(r, b.localFileResolver), nil
 	case provisioning.GitHubRepositoryType:
-		baseURL, err := url.Parse(b.urlProvider(namespace))
-		if err != nil {
-			return nil, fmt.Errorf("invalid base URL: %w", err)
-		}
-
 		// HACK: replace resource name placeholder in the webhook URL with the actual resource name
 		// In case the name was automatically generated by k8s
 		if r.Spec.GitHub.WebhookURL != "" {
 			r.Spec.GitHub.WebhookURL = strings.ReplaceAll(r.Spec.GitHub.WebhookURL, resourceNamePlaceholder, r.GetName())
 		}
 
-		dynamicClient, kinds, err := b.client.New(namespace)
-		if err != nil {
-			return nil, fmt.Errorf("failed to create dynamic client: %w", err)
-		}
-
-		parser := resources.NewParser(r, dynamicClient, kinds)
-		replicator := resources.NewReplicator(dynamicClient, parser, r.Spec.Folder)
-		linterFactory := lint.NewDashboardLinterFactory()
-
-		return repository.NewGitHub(ctx, r, b.ghFactory, baseURL, linterFactory, b.renderer, replicator), nil
+		return repository.NewGitHub(ctx, r, b.ghFactory), nil
 	case provisioning.S3RepositoryType:
 		return repository.NewS3(r), nil
 	default:
