+++ pkg/registry/apis/provisioning/repository/local.go
@@ -19,7 +19,6 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/validation/field"
-	"k8s.io/apiserver/pkg/registry/rest"
 
 	provisioning "github.com/grafana/grafana/pkg/apis/provisioning/v0alpha1"
 )
@@ -274,12 +273,6 @@ func (r *localRepository) History(ctx context.Context, logger *slog.Logger, path
 			Code:    http.StatusNotImplemented,
 		},
 	}
-}
-
-// Webhook implements provisioning.Repository.
-func (r *localRepository) Webhook(ctx context.Context, logger *slog.Logger, responder rest.Responder) http.HandlerFunc {
-	// webhooks are not supported with local
-	return nil
 }
 
 func (r *localRepository) AfterCreate(ctx context.Context, logger *slog.Logger) error {
