+++ pkg/registry/apis/provisioning/repository/unknown.go
@@ -9,7 +9,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/validation/field"
-	"k8s.io/apiserver/pkg/registry/rest"
 )
 
 var _ Repository = (*unknownRepository)(nil)
@@ -101,12 +100,6 @@ func (r *unknownRepository) History(ctx context.Context, logger *slog.Logger, pa
 			Code:    http.StatusNotImplemented,
 		},
 	}
-}
-
-// Webhook implements provisioning.Repository.
-func (r *unknownRepository) Webhook(ctx context.Context, logger *slog.Logger, responder rest.Responder) http.HandlerFunc {
-	// webhooks are not supported with local
-	return nil
 }
 
 func (r *unknownRepository) AfterCreate(ctx context.Context, logger *slog.Logger) error {
