+++ pkg/registry/apis/provisioning/types.go
@@ -4,8 +4,10 @@ import (
 	"context"
 
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository"
+	"github.com/grafana/grafana/pkg/registry/apis/provisioning/webhook"
 )
 
-type RepoGetter interface {
+type Getter interface {
 	GetRepository(ctx context.Context, name string) (repository.Repository, error)
+	GetWebhook(ctx context.Context, name string) (webhook.Webhook, error)
 }
